pyproject.toml
[project]
name = "robocoin-viewer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "h5py>=3.15.1",
    "lerobot>=0.4.3",
    "lz4>=4.4.5",
    "mcap>=1.3.1",
    "mcap-ros2-support>=0.5.7",
    "numpy>=2.2.6",
    "opencv-python>=4.13.0.90",
    "pandas>=2.3.3",
    "pyarrow>=23.0.0",
    "pydantic>=2.12.5",
    "rerun-sdk>=0.26.2",
    "rosbags[mcap]>=0.11.0",
    "streamlit>=1.53.1",
    "zstandard>=0.25.0",
    "zstd>=1.5.7.3",
]
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[[tool.uv.index]]
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
pythonpath = ["src"]
